Regression test: use UTF-8 string in test payload
See 18F/github-webhook-validator#2 for the motivation. Even though just
upgrading to github-webhook-validator v0.1.1 resolves the issue, I still
wanted to make sure that the tests in this package were UTF-8-clean. Learned a
surprising detail about UTF-8 Content-Length calculation in the process,
explained with a comment in the test data.