Fix HTTP 417 error when testing with Node v5.5.0
In Node v5.5.0, the HTTP server became more diligent about sending a HTTP 417
Expectation Failed response if the `Expect` header was not `100-continue`.
Since `RequestHelper.httpOptions` was creating requests with an empty `Expect`
header, and did not define a `checkExpectation` event handler, the server
returned 417 errors for these requests.

Also updated `RequestHelper.sendRequest` to use the status code text as the
error message if the error response body is empty.

https://nodejs.org/en/blog/release/v5.5.0/
https://github.com/nodejs/node/pull/4501
https://nodejs.org/api/http.html#http_event_checkexpectation
https://www.w3.org/Protocols/rfc2616/rfc2616-sec10.html#sec10.4.18
https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.20
https://www.w3.org/Protocols/rfc2616/rfc2616-sec8.html#sec8.2.3