Refactor finishBuild(), build() return statement
Per @msecret in #33, made the `finish` function more straightforward by
passing it an `err` argument, which `finishBuild()` will then use to reject or
resolve its promise. This means that calling `done` as the final handler is
straightforward, and easily removed in a later PR when we return the resulting
`Promise` directly.