Add repo name to dest dir for empty baseurl
The solution from #39 was incorrect. If there is a .18f-pages.yml file with an
empty `baseurl:` property, we should still generate it into a directory with
the same name as the repository. The web server configuration will need to
contain this directory as the document root for the target site.

For example, for 18F/handbook, the baseurl will be empty, the output directory
will be GENERATED_SITE_DIR/handbook, and the web server config will set
GENERATED_SITE_DIR/handbook as the root for handbook.18f.gov.