Sync class for backing up successful builds
This enables the `pages` container from 18F/knowledge-sharing-toolkit to sync
all of the existing sites locally upon startup. Should be straightforward to
replace the S3-specific bits with, say, rsync if that becomes a use case. (In
fact I implemented this, but decided it wasn't yet worth the complexity,
despite how easy it was.)